detail, settings

+++ app/src/main/java/com/zzy/champions/ui/detail/compose/BuildContent.kt
@@ -1,4 +1,4 @@
-package com.zzy.champions.ui.components
+package com.zzy.champions.ui.detail.compose
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -9,6 +9,8 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material3.Card
@@ -23,10 +25,15 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.zzy.champions.R
 import com.zzy.champions.data.model.ChampionBuild
+import com.zzy.champions.ui.components.ChampionBuildDialog
+import com.zzy.champions.ui.components.MenuDialog
+import com.zzy.champions.ui.components.TextDialog
 import com.zzy.champions.ui.theme.MyApplicationTheme
 
 @Composable
@@ -34,14 +41,14 @@ fun ChampionBuildScreen(
     modifier: Modifier = Modifier,
     builds: List<ChampionBuild>,
     onItemClick: (ChampionBuild) -> Unit,
-    onInsertItem: (Int, String, String) -> Unit,
+    onEditBuild: (ChampionBuild) -> Unit,
     onDeleteItem: (ChampionBuild) -> Unit,
 ) {
-    Column(
+    LazyColumn(
         modifier = modifier.padding(16.dp),
     ) {
-        builds.forEach {
-            ChampionBuildItem(cb = it, onClick = onItemClick, onEditClick = onInsertItem, onDeleteClick = onDeleteItem)
+        items(builds) {
+            ChampionBuildItem(cb = it, onClick = onItemClick, onEditClick = onEditBuild, onDeleteClick = onDeleteItem)
         }
     }
 }
@@ -51,10 +58,11 @@ fun ChampionBuildItem(
     modifier: Modifier = Modifier,
     cb: ChampionBuild,
     onClick: (ChampionBuild) -> Unit,
-    onEditClick: (Int, String, String) -> Unit,
+    onEditClick: (ChampionBuild) -> Unit,
     onDeleteClick: (ChampionBuild) -> Unit
 ) {
     var showMenu by remember { mutableStateOf(false) }
+    var showConfirmation by remember { mutableStateOf(false) }
     var showEditor by remember { mutableStateOf(false) }
 
     Column(modifier) {
@@ -108,19 +116,29 @@ fun ChampionBuildItem(
             },
             onDelete = {
                 showMenu = false
-                onDeleteClick(cb)
+                showConfirmation = true
             }
         )
+    }
+
+    if (showConfirmation) {
+        TextDialog(
+            onDismissRequest = { showConfirmation = false },
+            onPositiveButtonClick = {
+                onDeleteClick(cb)
+            },
+            content = {
+                Text(text = stringResource(id = R.string.do_you_want_to_delete_this_build))
+            })
     }
 
     if (showEditor) {
         ChampionBuildDialog(
             onDismissRequest = { showEditor = false },
-            initTitle = cb.nameOfBuild,
-            initContent = cb.url,
-            onOkClick = { title, content ->
+            build = cb,
+            onOkClick = { cb ->
                 showEditor = false
-                onEditClick(cb.id, title, content)
+                onEditClick(cb)
             }
         )
     }
@@ -137,7 +155,7 @@ fun PreviewBuilds() {
     MyApplicationTheme {
         ChampionBuildScreen(builds = list, onItemClick = {
 
-        }, onInsertItem = { a, b, c ->
+        }, onEditBuild = { c ->
 
         }, onDeleteItem = {
